// wooden-box/three-core.d.ts
export enum Combine { }
// ...
export enum Mapping { }
export const UVMapping: Mapping;
export const CubeReflectionMapping: Mapping;
export const CubeRefractionMapping: Mapping;
export const EquirectangularReflectionMapping: Mapping;
export const EquirectangularRefractionMapping: Mapping;
export const SphericalReflectionMapping: Mapping;
export const CubeUVReflectionMapping: Mapping;
export const CubeUVRefractionMapping: Mapping;

// Wrapping modes
export enum Wrapping { }
export const RepeatWrapping: Wrapping;
export const ClampToEdgeWrapping: Wrapping;
export const MirroredRepeatWrapping: Wrapping;

// Filters
export enum TextureFilter { }
export const NearestFilter: TextureFilter;
export const NearestMipMapNearestFilter: TextureFilter;
export const NearestMipMapLinearFilter: TextureFilter;
export const LinearFilter: TextureFilter;
export const LinearMipMapNearestFilter: TextureFilter;
export const LinearMipMapLinearFilter: TextureFilter;

// Data types
export enum TextureDataType { }
// ...
export enum PixelFormat { }
// ...
export enum TrianglesDrawModes { }
export const TrianglesDrawMode: TrianglesDrawModes;
export const TriangleStripDrawMode: TrianglesDrawModes;
export const TriangleFanDrawMode: TrianglesDrawModes;

// Texture Encodings
export enum TextureEncoding { }
// ...
export class EventDispatcher {

    /**
     * Creates eventDispatcher object. It needs to be call with '.call' to add the functionality to an object.
     */
    constructor();

    /**
     * Adds a listener to an event type.
     * @param type The type of event to listen to.
     * @param listener The function that gets called when the event is fired.
     */
    addEventListener(type: string, listener: (event: Event) => void ): void;

    /**
     * Checks if listener is added to an event type.
     * @param type The type of event to listen to.
     * @param listener The function that gets called when the event is fired.
     */
    hasEventListener(type: string, listener: (event: Event) => void): boolean;

    /**
     * Removes a listener from an event type.
     * @param type The type of the listener that gets removed.
     * @param listener The listener function that gets removed.
     */
    removeEventListener(type: string, listener: (event: Event) => void): void;

    /**
     * Fire an event type.
     * @param type The type of event that gets fired.
     */
    dispatchEvent(event: { type: string; [attachment: string]: any; }): void;
}
// ...
export interface Intersection {
    distance: number;
    distanceToRay?: number;
    point: Vector3;
    index?: number;
    face?: Face3 | null;
    faceIndex?: number;
    object: Object3D;
    uv?: Vector2;
}
// ...
export class Loader {
    constructor();

    /**
     * Will be called when load starts.
     * The default is a function with empty body.
     */
    onLoadStart: () => void;

    /**
     * Will be called while load progresses.
     * The default is a function with empty body.
     */
    onLoadProgress: () => void;

    /**
     * Will be called when load completes.
     * The default is a function with empty body.
     */
    onLoadComplete: () => void;

    /**
     * default — null.
     * If set, assigns the crossOrigin attribute of the image to the value of crossOrigin, prior to starting the load.
     */
    crossOrigin: string;

    /**
     * @deprecated Use THREE.LoaderUtils.extractUrlBase() instead.
     */
    extractUrlBase(url: string): string;
    initMaterials(materials: Material[], texturePath: string): Material[];
    createMaterial(m: Material, texturePath: string, crossOrigin?: string): boolean;

    static Handlers: LoaderHandler;
}

/**
* Interface for all loaders
* CompressedTextureLoader don't extends Loader class, but have load method
*/
export interface AnyLoader {
  load(url: string, onLoad?: (result: any) => void, onProgress?: (event: ProgressEvent) => void, onError?: (event: ErrorEvent) => void): any;
}

export interface LoaderHandler {
    handlers: (RegExp | AnyLoader)[];

    add(regex: RegExp, loader: AnyLoader): void;
    get(file: string): AnyLoader | null;
}
// ...
export class ImageLoader {
    constructor(manager?: LoadingManager);

    manager: LoadingManager;
    crossOrigin: string;
    withCredentials: string;
    path: string;

    /**
     * Begin loading from url
     * @param url
     */
    load(url: string, onLoad?: (image: HTMLImageElement) => void, onProgress?: (event: ProgressEvent) => void, onError?: (event: ErrorEvent) => void): HTMLImageElement;
    setCrossOrigin(crossOrigin: string): ImageLoader;
    setWithCredentials(value: string): ImageLoader;
    setPath(value: string): ImageLoader;
}

/**
 * A loader for loading objects in JSON format.
 */
export class JSONLoader extends Loader {
    constructor(manager?: LoadingManager);

    manager: LoadingManager;
    withCredentials: boolean;

    load(url: string, onLoad?: (geometry: Geometry, materials: Material[]) => void, onProgress?: (event: ProgressEvent) => void, onError?: (event: ErrorEvent) => void): void;
    setTexturePath( value: string ): void;
    parse(json: any, texturePath?: string): { geometry: Geometry; materials?: Material[] };
}

/**
 * Handles and keeps track of loaded and pending data.
 */
export class LoadingManager {
    constructor(onLoad?: () => void, onProgress?: (url: string, loaded: number, total: number) => void, onError?: () => void);

    onStart?: (url: string, loaded: number, total: number) => void;

    /**
     * Will be called when load starts.
     * The default is a function with empty body.
     */
    onLoad: () => void;

    /**
     * Will be called while load progresses.
     * The default is a function with empty body.
     */
    onProgress: (item: any, loaded: number, total: number) => void;

    /**
     * Will be called when each element in the scene completes loading.
     * The default is a function with empty body.
     */
    onError: (url: string) => void;

    /**
     * If provided, the callback will be passed each resource URL before a request is sent.
     * The callback may return the original URL, or a new URL to override loading behavior.
     * This behavior can be used to load assets from .ZIP files, drag-and-drop APIs, and Data URIs.
     * @param callback URL modifier callback. Called with url argument, and must return resolvedURL.
     */
    setURLModifier(callback?: (url: string) => string): void;

    /**
     * Given a URL, uses the URL modifier callback (if any) and returns a resolved URL.
     * If no URL modifier is set, returns the original URL.
     * @param url the url to load
     */
    resolveURL(url: string): string;

    itemStart(url: string): void;
    itemEnd(url: string): void;
    itemError(url: string): void;
}
// ...
export interface MeshBasicMaterialParameters extends MaterialParameters {
    color?: Color | string | number;
    opacity?: number;
    map?: Texture;
    aoMap?: Texture;
    aoMapIntensity?: number;
    specularMap?: Texture;
    alphaMap?: Texture;
    envMap?: Texture;
    combine?: Combine;
    reflectivity?: number;
    refractionRatio?: number;
    wireframe?: boolean;
    wireframeLinewidth?: number;
    wireframeLinecap?: string;
    wireframeLinejoin?: string;
    skinning?: boolean;
    morphTargets?: boolean;
}

export class MeshBasicMaterial extends Material {
    constructor(parameters?: MeshBasicMaterialParameters);

    color: Color;
    map: Texture | null;
    aoMap: Texture | null;
    aoMapIntensity: number;
    specularMap: Texture | null;
    alphaMap: Texture | null;
    envMap: Texture | null;
    combine: Combine;
    reflectivity: number;
    refractionRatio: number;
    wireframe: boolean;
    wireframeLinewidth: number;
    wireframeLinecap: string;
    wireframeLinejoin: string;
    skinning: boolean;
    morphTargets: boolean;

    setValues(parameters: MeshBasicMaterialParameters): void;
}
// ...
export interface HSL {
    h: number;
    s: number;
    l: number;
}

/**
 * Represents a color. See also {@link ColorUtils}.
 *
 * @example
 * var color = new THREE.Color( 0xff0000 );
 *
 * @see <a href="https://github.com/mrdoob/three.js/blob/master/src/math/Color.js">src/math/Color.js</a>
 */
export class Color {
    constructor(color?: Color|string|number);
    constructor(r: number, g: number, b: number);

    /**
     * Red channel value between 0 and 1. Default is 1.
     */
    r: number;

    /**
     * Green channel value between 0 and 1. Default is 1.
     */
    g: number;

    /**
     * Blue channel value between 0 and 1. Default is 1.
     */
    b: number;

    set(color: Color): Color;
    set(color: number): Color;
    set(color: string): Color;
    setScalar(scalar: number): Color;
    setHex(hex: number): Color;

    /**
     * Sets this color from RGB values.
     * @param r Red channel value between 0 and 1.
     * @param g Green channel value between 0 and 1.
     * @param b Blue channel value between 0 and 1.
     */
    setRGB(r: number, g: number, b: number): Color;

    /**
     * Sets this color from HSL values.
     * Based on MochiKit implementation by Bob Ippolito.
     *
     * @param h Hue channel value between 0 and 1.
     * @param s Saturation value channel between 0 and 1.
     * @param l Value channel value between 0 and 1.
     */
    setHSL(h: number, s: number, l: number): Color;

    /**
     * Sets this color from a CSS context style string.
     * @param contextStyle Color in CSS context style format.
     */
    setStyle(style: string): Color;

    /**
     * Clones this color.
     */
    clone(): this;

    /**
     * Copies given color.
     * @param color Color to copy.
     */
    copy(color: Color): this;

    /**
     * Copies given color making conversion from gamma to linear space.
     * @param color Color to copy.
     */
    copyGammaToLinear(color: Color, gammaFactor?: number): Color;

    /**
     * Copies given color making conversion from linear to gamma space.
     * @param color Color to copy.
     */
    copyLinearToGamma(color: Color, gammaFactor?: number): Color;

    /**
     * Converts this color from gamma to linear space.
     */
    convertGammaToLinear(): Color;

    /**
     * Converts this color from linear to gamma space.
     */
    convertLinearToGamma(): Color;

    /**
     * Returns the hexadecimal value of this color.
     */
    getHex(): number;

    /**
     * Returns the string formated hexadecimal value of this color.
     */
    getHexString(): string;

    getHSL(target: HSL): HSL;

    /**
     * Returns the value of this color in CSS context style.
     * Example: rgb(r, g, b)
     */
    getStyle(): string;

    offsetHSL(h: number, s: number, l: number): this;

    add(color: Color): this;
    addColors(color1: Color, color2: Color): this;
    addScalar(s: number): this;
    sub(color: Color): this;
    multiply(color: Color): this;
    multiplyScalar(s: number): this;
    lerp(color: Color, alpha: number): this;
    lerpHSL(color: Color, alpha: number): this;
    equals(color: Color): boolean;
    fromArray(rgb: number[], offset?: number): this;
    toArray(array?: number[], offset?: number): number[];
    toArray(xyz: ArrayLike<number>, offset?: number): ArrayLike<number>;
}
// ...
export class Mesh extends Object3D {
    constructor(geometry?: Geometry | BufferGeometry, material?: Material | Material[]);

    geometry: Geometry | BufferGeometry;
    material: Material | Material[];
    drawMode: TrianglesDrawModes;
    morphTargetInfluences?: number[];
    morphTargetDictionary?: { [key: string]: number; };
    isMesh: true;
    type: string;

    setDrawMode(drawMode: TrianglesDrawModes): void;
    updateMorphTargets(): void;
    raycast(raycaster: Raycaster, intersects: Intersection[]): void;
    copy(source: this, recursive?: boolean): this;
}
// ...
export class Texture extends EventDispatcher {
    constructor(
        image?: HTMLImageElement | HTMLCanvasElement | HTMLVideoElement,
        mapping?: Mapping,
        wrapS?: Wrapping,
        wrapT?: Wrapping,
        magFilter?: TextureFilter,
        minFilter?: TextureFilter,
        format?: PixelFormat,
        type?: TextureDataType,
        anisotropy?: number,
        encoding?: TextureEncoding
        );

    id: number;
    uuid: string;
    name: string;
    sourceFile: string;
    image: any; // HTMLImageElement or ImageData or { width: number, height: number } in some children;
    mipmaps: ImageData[];
    mapping: Mapping;
    wrapS: Wrapping;
    wrapT: Wrapping;
    magFilter: TextureFilter;
    minFilter: TextureFilter;
    anisotropy: number;
    format: PixelFormat;
    type: TextureDataType;
    offset: Vector2;
    repeat: Vector2;
    center: Vector2;
    rotation: number;
    generateMipmaps: boolean;
    premultiplyAlpha: boolean;
    flipY: boolean;
    unpackAlignment: number;
    encoding: TextureEncoding;
    version: number;
    needsUpdate: boolean;
    onUpdate: () => void;
    static DEFAULT_IMAGE: any;
    static DEFAULT_MAPPING: any;

    clone(): this;
    copy(source: Texture): this;
    toJSON(meta: any): any;
    dispose(): void;
    transformUv(uv: Vector): void;
}

export class DepthTexture extends Texture {
    constructor(
        width: number,
        heighht: number,
        type?: TextureDataType,
        mapping?: Mapping,
        wrapS?: Wrapping,
        wrapT?: Wrapping,
        magFilter?: TextureFilter,
        minFilter?: TextureFilter,
        anisotropy?: number
    );

    image: { width: number, height: number };
}

export class CanvasTexture extends Texture {
    constructor(
        canvas: HTMLImageElement | HTMLCanvasElement | HTMLVideoElement,
        mapping?: Mapping,
        wrapS?: Wrapping,
        wrapT?: Wrapping,
        magFilter?: TextureFilter,
        minFilter?: TextureFilter,
        format?: PixelFormat,
        type?: TextureDataType,
        anisotropy?: number
    );
}